Reverse Input direction when camera in near mode
Because the camera rotates 180 deg in near mode, the left and right keys
need to be reversed in near mode. However, if the player is already holding
one of those keys, that key should not be reversed until the key is released
and then re-pressed. It's a bit complex, but it's the way that Metal Gear
Solid handled the issue.